Some fixes

+++ build.ps1
@@ -1,4 +1,5 @@
 param(
+    [switch] $CleanDist = $true,
     [switch] $RebuildShell = $false,
     [switch] $RebuildBuildImages = $false
 )
@@ -20,7 +21,7 @@ if (($NgxImageExists -and $RebuildBuildImages) -or (-not $NgxImageExists)) {
     if ($NgxImageExists) {
         docker rmi ${NgxImage} --force
     }
-    docker build -f ./docker/NgBuild.Dockerfile -t ${NgxImage} .
+    docker build -f ./docker/NgxBuild.Dockerfile -t ${NgxImage} .
 }
 
 function Ng-Build {
@@ -39,16 +40,16 @@ function Ng-Build {
     docker cp . ${NgContainer}:/usr/src/app/${Dir}
     if (-Not $IsShell) {
         docker exec ${NgContainer} /bin/sh -c "mkdir ${NgShell}"
-        docker cp ../${NgShell}/dist ${NgContainer}:/usr/src/app/${NgShell}
-        docker cp ../iqs-libs-clientshell-angular ${NgContainer}:/usr/src/app
+        docker cp ../${NgShell}/dist ${NgContainer}:/usr/src/app/${NgShell}/dist
+        docker cp ../iqs-libs-unsupported-angular ${NgContainer}:/usr/src/app
     }
     docker exec ${NgContainer} /bin/sh -c "cd ${Dir} && gulp build"
-    if (Test-Path -Path ".\dist") {
-		Remove-Item -LiteralPath "dist" -Recurse -Force
-    }
     if (-Not $IsShell) {
         docker cp ${NgContainer}:/usr/src/app/${Dir}/dist ../dist/${Dist}
     } else {
+        if (Test-Path -Path ".\dist") {
+            Remove-Item -LiteralPath "dist" -Recurse -Force
+        }
         docker cp ${NgContainer}:/usr/src/app/${Dir}/dist .
     }
     docker stop $NgContainer
@@ -67,7 +68,7 @@ function Ngx-Build {
     if (docker ps -q -a -f name=$NgxContainer) {
         docker rm $NgxContainer --force
     }
-    docker run -t -d --name $NgxContainer $NgxImage
+    docker run --memory 2048m -t -d --name $NgxContainer $NgxImage
     tar -cf "${Dir}.tar" --exclude=node_modules --exclude=.git --exclude=coverage --exclude=dist --exclude="${Dir}.tar" *
     docker cp ./$Dir.tar ${NgxContainer}:/usr/src/app
     Remove-Item -LiteralPath "${Dir}.tar"
@@ -87,11 +88,15 @@ if ($RebuildShell) {
 }
 
 if (Test-Path -Path ".\dist") {
-    Remove-Item -LiteralPath "dist" -Recurse -Force
+    if ($CleanDist) {
+        Remove-Item -LiteralPath "dist" -Recurse -Force
+    }
 }
 New-Item -itemtype "directory" -Path . -Name "dist" -Force
-New-Item .\dist\index.html
-Set-Content .\dist\index.html '<head><meta http-equiv="refresh" content="1;URL=/home/index.html" /></head>'
+if (!(Test-Path -Path ".\dist\index.html" -PathType Leaf)) {
+    New-Item .\dist\index.html
+    Set-Content .\dist\index.html '<head><meta http-equiv="refresh" content="1;URL=/home/index.html" /></head>'
+}
 
 $Dists | ForEach-Object {
     if ($_.type -eq 'ng') {
@@ -104,4 +109,4 @@ $Dists | ForEach-Object {
 if (docker images ${ClientsImage} -q) {
     docker rmi ${ClientsImage} --force
 }
-docker build -f ./docker/Nginx.Dockerfile -t ${ClientsImage} .
+docker build --no-cache -f ./docker/Nginx.Dockerfile -t ${ClientsImage} .
